% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleCervus.R
\name{ResampleCervus}
\alias{ResampleCervus}
\title{Parentage from Resampling}
\usage{
ResampleCervus(
  CervusCL_directory,
  Data_directory,
  Genotype_file,
  Genotype_file.ID_in_column,
  Genotype_file.First_allele_in_column,
  Parent_file,
  Offspring_file,
  Simulate.offspring = 10000,
  Simulate.Prop.sampled = 0.8,
  Simulate.Prop.mistyped = 0.001,
  Simulate.Minimum_typed_loci = "defaulted",
  Genotype_file.HeaderRow = 1,
  Parent_file.HeaderRow = 1,
  Parent_file.ID_in_column = 1,
  Offspring_file.HeaderRow = 1,
  Offspring_file.ID_in_column = 1,
  Add_sex = FALSE,
  Add_sex.Gender_file = NULL,
  Add_sex.Simulate.Prop.father.sampled = 0.8,
  resample.prop = 0.8,
  resample.nsim = 100,
  ignore.stdout = TRUE,
  ignore.hasexecuted = FALSE
)
}
\arguments{
\item{CervusCL_directory}{path of directory that contains "\code{CervusCL.exe}"}

\item{Data_directory}{path of directory that contains data files, which include
genotypic data (as well as individual ID), parent ID, offspring ID, and gender
information (optional). All files should be presented as CSV format.}

\item{Genotype_file}{name of genotypic data file. Genotype information should be
organized as a format that can be read by CervusCL. Each row is an individual (except
the header row), and every two columns is a loci (except the ID column). The ID
column should be previous to loci information.}

\item{Genotype_file.ID_in_column}{index of ID column for genotypic data file.}

\item{Genotype_file.First_allele_in_column}{index of the start column of loci information
for genotypic data file.}

\item{Parent_file}{name of parent ID file. Only the first column is used unless
\code{Parent_file.ID_in_column} is modified. It contains all IDs for parental
individuals.}

\item{Offspring_file}{name of offspring ID file. Only the first column is used
unless \code{Offspring_file.ID_in_column} is modified. It contains all IDs for
offspring individuals.}

\item{Simulate.offspring}{number of simulated offspring. Defaulted is 10000, as
consistent with Cervus. Change of this value is not suggested.}

\item{Simulate.Prop.sampled}{proportion of sampled individuals in the population.
Default is 0.8 but its actual value may vary among different populations although
estimation of this value is challenging.}

\item{Simulate.Prop.mistyped}{genotyping error rate which may due to alleleic dropout,
stochastic typing error, mutation, etc. Estimation of this value is also challenging.}

\item{Simulate.Minimum_typed_loci}{minimum typed loci. Individuals with typed loci
lower than this value will be discarded. Defaulted is the half of available loci, as
consistent with Cervus.}

\item{Genotype_file.HeaderRow}{whether to contain header row in genotypic data file.}

\item{Parent_file.HeaderRow}{whether to contain header to in parental ID file.}

\item{Parent_file.ID_in_column}{index of ID column in parental ID file.}

\item{Offspring_file.HeaderRow}{whether to contain header to in offspring ID file.}

\item{Offspring_file.ID_in_column}{index of ID column in offspring ID file.}

\item{Add_sex}{whether to add gender information.}

\item{Add_sex.Gender_file}{name of gender data file. The file should contain only
one column with each row showing gender information for an individual. The order of
row should be consistent with that in genotypic file. Gender can be donated by any
numbers and/or characters. For example, use 1 to donate males and 2 donating females.
Note that \code{1 < 2}, and the smaller level is considered as males while higher level
for females.}

\item{Add_sex.Simulate.Prop.father.sampled}{when using gender information,
\code{Simulate.Prop.sampled} is used to donate sampling rate for mothers while
this parameter is for fathers.}

\item{resample.prop}{proportion of resampled individuals for each replicate.}

\item{resample.nsim}{number of replicates.}

\item{ignore.stdout}{inhibit outputs when executing CervusCL.}

\item{ignore.hasexecuted}{ignore files that has been executed. Sometimes, you may
finish CervusCL algorithm for part of the replicates while the remaining is not
executed yet. In the next time, you can continue your work by setting this parameter to
\code{TRUE} to perform parentage for the remaining files.}
}
\value{
no returned values.
}
\description{
\code{ResampleCervus} is used to perform parentage analysis by CervusCL using
resampling approach.
}
\examples{
\dontrun{
CervusCL_directory <- "D:/xxx/Cervus/Cervus CL"
Data_directory <- "D:/xxx"
Genotype_file <- "Genotype_data.csv"
Genotype_file.ID_in_column <- 1
Genotype_file.First_allele_in_column <- 2
Parent_file <- "Parent_ID_data.csv"
Offspring_file <- "Offspring_ID_data.csv"
Gender_file <- "Gender_data.csv"

ResampleCervus(CervusCL_directory, Data_directory,
               Genotype_file, Genotype_file.ID_in_column, Genotype_file.First_allele_in_column,
               Parent_file, Offspring_file, ignore.hasexecuted = TRUE,
               Add_sex = TRUE, Add_sex.Gender_file = Gender_file, resample.prop = 0.8)
}
}
\references{
Marshall, T.C., Slate, J., Kruuk, L.E.B. & Pemberton, J.M. (1998). Statistical confidence for likelihood-based paternity inference in natural populations. Molecular Ecology, 7, 639-655.
}
\author{
Hao-Ran Wu
}
