% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{ResampleResult}
\alias{ResampleResult}
\title{Summarize Parentage Result}
\usage{
ResampleResult(filepath, Add_sex = FALSE)
}
\arguments{
\item{filepath}{path of directory that contains parentage result files. It is the
same directory assigned by parameter \code{Data_directory} in function "\code{ResampleCervus}".}

\item{Add_sex}{whether sex data is used in parentage analysis.}
}
\value{
a \code{data.frame} of four columns. The first two columns are parents. The
third column shows information about offspring. The last column is per–triad support value,
which is defined as the number of replicates that support a triad divided by the
number of samples that include the individuals of the triad.
}
\description{
\code{ResampleResult} reports parentage result from summarizing parent-offspring
pairs that occur in each replicate.
}
\examples{
\dontrun{
Data_directory <- "D:/xxx"
result <- ResampleResult(Data_directory, Add_sex = TRUE)
print(result)
}
}
\author{
Hao-Ran Wu
}
